Implement more features and feed it into a KNN model
I've reworked main.py. All the numerical features would be transformed using Scikit-learn's StandardScaler function
for consistency.

The "genre" feature was transformed using a MultiLabelBinarizer instead of a OneHotEncoder

The "title" feature was transformed using the Word Embedding technique. I used a pretrained model from the spaCy library.
A seperate KNN model was feed of this feature to act as a search engine.

The "overview" or description feature would be implemented later.